**final update** see description
it never updated that we changed the casing of the  player movement script, so we changed it from the definition. if player movement does not work you may need to make this change so they match.

if it cannot look up and down, the "player model goes here" object needs the main camera to be added to the script.